refactor: generalize insert method
